+++ dotnet/Api.Music/src/Api.Music.csproj
