+++ dotnet/Api.Music/src/Controllers/Models/BaseUpdateRequest.cs
