+++ dotnet/Api.Music/src/Controllers/Models/MusicAddRequest.cs
