+++ dotnet/Api.Music/src/Controllers/Models/MusicResponse.cs
