+++ dotnet/Api.Music/src/Controllers/Models/MusicSearchRequest.cs
