+++ dotnet/Api.Music/src/Controllers/Models/MusicUpdateFavoriteRequest.cs
