+++ dotnet/Api.Music/src/Controllers/Models/MusicUpdateRequest.cs
