+++ dotnet/Api.Music/src/Properties/launchSettings.json
