+++ dotnet/Api.Music/src/Repositories/Documents/BaseDocument.cs
