+++ dotnet/Api.Music/src/Repositories/Documents/MusicDocument.cs
