+++ dotnet/Api.Music/src/Repositories/IMusicRepository.cs
