+++ dotnet/Api.Music/src/Repositories/MusicExtension.cs
