+++ dotnet/Api.Music/src/Repositories/Settings/IMongoDBSetting.cs
