+++ dotnet/Api.Music/src/Repositories/Settings/MongoDBSetting.cs
