+++ dotnet/Api.Music/src/Services/IMusicService.cs
