+++ dotnet/Api.Music/src/Services/MusicMongoMapper.cs
