+++ dotnet/Api.Music/src/Services/MusicService.cs
