+++ dotnet/Api.Music/src/Startup.cs
