+++ dotnet/Api.Music/src/appsettings.Development.json
