+++ dotnet/Api.Music/src/appsettings.json
